requirements.txt
# Todas tus dependencias existentes mantenidas exactamente como las proporcionaste
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
altgraph==0.17
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
bcrypt==4.2.1
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
coreapi==2.3.3
coreschema==0.0.4
croniter==6.0.0
cryptography==44.0.0
debtcollector==3.0.0
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.17
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
dogpile.cache==1.3.3
eventlet==0.38.2
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
filetype==1.2.0
frozenlist==1.5.0
future==0.18.2
futurist==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
iso8601==2.1.0
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keystoneauth1==5.9.1
keystonemiddleware==10.8.0
kombu==5.4.2
langchain==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.2.6
logutils==0.3.5
macholib==1.15.2
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistral==19.0.0
mistral-lib==3.0.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.58.1
orjson==3.10.12
os-service-types==1.7.0
oslo.cache==3.9.0
oslo.concurrency==6.2.0
oslo.config==9.7.0
oslo.context==5.7.0
oslo.db==17.1.0
oslo.i18n==6.5.0
oslo.log==6.2.0
oslo.messaging==15.0.0
oslo.metrics==0.10.0
oslo.middleware==6.3.0
oslo.policy==4.5.0
oslo.serialization==5.6.0
oslo.service==3.6.0
oslo.utils==8.0.0
osprofiler==4.2.0
packaging==24.2
paramiko==3.5.0
parso==0.8.4
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.0
pecan==1.5.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
ply==3.11
prettytable==3.12.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycadf==4.0.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-keystoneclient==5.5.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.22.3
rsa==4.9
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
stack-data==0.6.3
statsd==4.0.1
stevedore==5.4.0
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tiktoken==0.8.0
tooz==6.3.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
whitenoise==6.8.2
wrapt==1.17.0
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
yarl==1.18.3
zipp==3.21.0
